# perf(array): optimize performance of literal expression