--- a/src/common/src/array/iterator.rs
+++ b/src/common/src/array/iterator.rs
@@ -50,6 +50,38 @@ impl<'a, A: Array> Iterator for ArrayIterator<'a, A> {
     }
 }
 
+pub struct ArrayRawIter<'a, A: Array> {
+    data: &'a A,
+    pos: usize,
+}
+
+impl<'a, A: Array> ArrayRawIter<'a, A> {
+    pub fn new(data: &'a A) -> Self {
+        Self { data, pos: 0 }
+    }
+}
+
+unsafe impl<'a, A: Array> TrustedLen for ArrayRawIter<'a, A> {}
+
+impl<'a, A: Array> Iterator for ArrayRawIter<'a, A> {
+    type Item = A::RefItem<'a>;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        if self.pos >= self.data.len() {
+            None
+        } else {
+            let item = self.data.value_at_raw(self.pos);
+            self.pos += 1;
+            Some(item)
+        }
+    }
+
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        let size = self.data.len() - self.pos;
+        (size, Some(size))
+    }
+}
+
 pub struct ArrayImplIterator<'a> {
     data: &'a ArrayImpl,
     pos: usize,