--- a/src/common/src/buffer/bitmap.rs
+++ b/src/common/src/buffer/bitmap.rs
@@ -404,7 +404,8 @@ impl Not for Bitmap {
 
 impl FromIterator<bool> for Bitmap {
     fn from_iter<T: IntoIterator<Item = bool>>(iter: T) -> Self {
-        let mut builder = BitmapBuilder::default();
+        let iter = iter.into_iter();
+        let mut builder = BitmapBuilder::with_capacity(iter.size_hint().0);
         for b in iter {
             builder.append(b);
         }
@@ -414,7 +415,8 @@ impl FromIterator<bool> for Bitmap {
 
 impl FromIterator<Option<bool>> for Bitmap {
     fn from_iter<T: IntoIterator<Item = Option<bool>>>(iter: T) -> Self {
-        let mut builder = BitmapBuilder::default();
+        let iter = iter.into_iter();
+        let mut builder = BitmapBuilder::with_capacity(iter.size_hint().0);
         for b in iter {
             builder.append(b.unwrap_or(false));
         }