I hereby agree to the terms of the [Singularity Data, Inc. Contributor License Agreement](https://gist.github.com/skyzh/0663682a70b0edde7ae991492f2314cb#file-s9y_cla).

## What's changed and what's your intention?

- introduce `Array::raw_iter` to efficiently iterate over raw values regardless of null.
- introduce `ArrayBuilder::append_n` to efficiently append the same value multiple times.
- remove the `head` from `Bitmap`

With these optimizations, the evaluation time of literal expressions dropped from 2.2us to 150ns.

This PR also adds an optimized version of `i32 + i32` using raw iterator in the bench, reducing time from 3.3us to 350us.

## Checklist

- [x] I have written necessary rustdoc comments
- [x] I have added necessary unit tests and integration tests
- [x] All checks passed in `./risedev check` (or alias, `./risedev c`)

## Refer to a related PR or issue link (optional)
#6868