@st1page 2022-12-13
> 3.3us to 350us

350ns?