@wangrunji0408 2022-12-13
> > 3.3us to 350us
> 
> 350ns?

Sure. Thanks for pointing it out. 🥵