@wangrunji0408 2022-12-13
> How should use the append_n in executor (e.g. build a chunk) or is it only for bench? Usually we do not how many values are the same before.

It is a method in `ArrayBuilder` trait. There's also a new method `append_datum_n` in `ArrayBuilderImpl`. So they should be generally available. But I think the most use case would still be evaluating literal expressions.

> I saw https://github.com/risingwavelabs/risingwave/pull/6856 has mentioned some potential improvements. Are all kinds of improvements has been includeded in this pr or there is more?

Yes, that's all.