@jon-chuang 2022-12-13
Although we demonstrated that we can improve the i32 add performance possibly by vectorizing each of the steps (null check, arithmetic and check overflow), can I ask whether any benchmark demonstrates improvement due to the use of `append_n`? 

I guess the intention is simply to provide an additional method that we can benchmark the usage of in the future?