@wangrunji0408 2022-12-13
> Although we demonstrated that we can improve the i32 add performance possibly by vectorizing each of the steps (null check, arithmetic and check overflow), can I ask whether any benchmark demonstrates improvement due to the use of `append_n`?

Sure. The bench "expr/constant" (literal expression). Previously it could only append the same value one by one. Now it can append them all at once.